fix: address kyli87 review feedback on metrics tests
- Simplify module doc comment to focus on I/O expectations, remove
  implementation-specific references (LogSegment::try_new, for_snapshot)
- Remove implementation references from test doc comments
- Use unwrap_committed() + post_commit_snapshot() to chain snapshots
  instead of rebuilding from scratch in insert_rows and setup helpers
- Move setup_in_memory_table and insert_rows helpers to main.rs for
  shared access across test modules
- Reuse setup_in_memory_table in scan.rs instead of inline setup
- Fix scan.rs scenario numbering (was 10, now unnumbered)
